Export service layer from api-server
