+++ apps/api-server/nodemon.json
@@ -2,5 +2,5 @@
   "watch": ["src"],
   "ignore": ["src/**/*.test.ts"],
   "ext": "ts,js,json",
-  "exec": "ts-eager src/index.ts"
+  "exec": "ts-eager src/main.ts"
 }
