+++ apps/api-server/package.json
@@ -7,7 +7,7 @@
     "build": "exit 0",
     "dev-resources:start": "s3rver -p 9000 -d ./s3_data --no-vhost-buckets --configure-bucket resources",
     "migrate": "ts-eager src/migrate.ts",
-    "start": "ts-eager src/index.ts",
+    "start": "ts-eager src/main.ts",
     "watch": "nodemon"
   },
   "license": "UNLICENSED",
